Making repo more release friendly
