--- ctf_builder/__init__.py
@@ -1,1 +1,0 @@
-from .schema import *
